fix: wineprefix folder detection

+++ plus/opt/Future_Pinball/fpinball_launcher.sh
@@ -22,6 +22,7 @@ RESOLUTION="${2}"
 RATIO="${3}"
 P1GUID="${4}"
 P1NAME="${5}"
+WINE=wine-lutris
 
 ################################################################################
 
@@ -31,8 +32,6 @@ FP_DIR='/opt/Future_Pinball'
 FP_STDIR="$HOME/configs/Future Pinball"
 FP_SVDIR="$HOME/../saves/Future Pinball"
 RES_START="$(batocera-resolution currentMode)"
-OPT_WINE="$(stat -t '/opt/Wine/wine-lutris/share/wine/wine.inf' | awk '{print $12}')"
-SVDIR_VERSION="$(cat -e "${FP_SVDIR}/wine/.update-timestamp" | cut -d '^' -f 1)"
 
 ################################################################################
 
@@ -115,19 +114,20 @@ function createFolders()
 
 function applyConfig()
 {
-    ### Aplica a config padrão
-    wine-lutris regedit "${FP_DIR}/emulator/config.reg"
-
     ### install deps for future pinball
-    batocera-load-screen -t 600 -i '/opt/Wine/wine.png' &
+    export WINE_LOADINGSCREEN=0
+    batocera-load-screen -t 600 -i '/opt/Future_Pinball/fpinball.jpg' &
+
+    ### Aplica a config padrão
+    "$WINE" regedit "${FP_DIR}/emulator/config.reg"
 
     cp -rf "${FP_DIR}/dll/"* "$WINEPREFIX/drive_c/windows/syswow64" && cp -rf "${FP_DIR}/dll/"* "$WINEPREFIX/drive_c/windows/system32"
 
     DLLS='dispex.dll jscript.dll scrobj.dll scrrun.dll vbscript.dll wshcon.dll wshext.dll'
 
     for i in $DLLS; do
-        wine-lutris regsvr32 "$WINEPREFIX/drive_c/windows/syswow64/$i" 2>&1&> /dev/null
-        wine-lutris regsvr32 "$WINEPREFIX/drive_c/windows/system32/$i" 2>&1&> /dev/null
+        "$WINE" regsvr32 "$WINEPREFIX/drive_c/windows/syswow64/$i" 2>&1&> /dev/null
+        "$WINE" regsvr32 "$WINEPREFIX/drive_c/windows/system32/$i" 2>&1&> /dev/null
         echo "$i Registrada com sucesso!"
     done
 
@@ -142,8 +142,13 @@ function applyConfig()
 
 ### CHECK FOLDERS
 
-if [ "${OPT_WINE}" != "${SVDIR_VERSION}" ] && [ -e "${FP_SVDIR}/wine" ]; then # se a versão do wine mudou
-    rm -r "${FP_SVDIR}/wine"
+if [ -e "${FP_SVDIR}/wine/.update-timestamp" ]; then # if wine version has changed
+    SVDIR_VERSION="$(cat -e "${FP_SVDIR}/wine/.update-timestamp" | cut -d '^' -f 1)"
+    OPT_WINE="$(stat -t "/opt/Wine/${WINE}/share/wine/wine.inf" | awk '{print $12}')"
+
+    if [ "${OPT_WINE}" != "${SVDIR_VERSION}" ] && [ -e "${FP_SVDIR}/wine" ]; then
+        rm -r "${FP_SVDIR}/wine"
+    fi
 fi
 
 if [ ! "$(ls -A "${FP_STDIR}" 2> /dev/null)" ] || [ ! "$(ls -A "${FP_SVDIR}" 2> /dev/null)" ]; then # configura do zero
@@ -236,9 +241,9 @@ JOGO="$(echo "Z:${JOGO}" | sed -e 's#/#\\#g')"
 export VIRTUAL_DESKTOP=1
 
 if [ "${JOGO}" == '' ]; then
-    wine-lutris "${FP_STDIR}/Future Pinball.exe"
+    "$WINE" "${FP_STDIR}/Future Pinball.exe"
 else
-    wine-lutris "${FP_STDIR}/Future Pinball.exe" /open "${JOGO}" /play /exit
+    "$WINE" "${FP_STDIR}/Future Pinball.exe" /open "${JOGO}" /play /exit
 fi
 
 ################################################################################
